+++ shake/wrappers.py
@@ -4,8 +4,6 @@
     --------------------------
 
 """
-from babel import Locale
-from pytz import timezone, UTC
 from werkzeug.utils import cached_property
 from werkzeug.wrappers import Request as BaseRequest
 from werkzeug.wrappers import Response as BaseResponse
@@ -64,11 +62,6 @@ class Request(BaseRequest):
     # Set by the application
     max_form_memory_size = 0
 
-    # The request locale
-    locale = None
-
-    # The request timezone (UTC if not defined)
-    tzinfo = None
     
     @property
     def is_get(self):
@@ -111,49 +104,6 @@ class Request(BaseRequest):
         if not data:
             return SecureCookie(secret_key=secret_key)
         return SecureCookie.unserialize(data, secret_key)
-
-    
-    def get_timezone(self, default=UTC):
-        """Returns the timezone that should be used for this request as a
-        `DstTzInfo` instance.
-
-        Tries the following in order:
-        - an attribute called `'tzinfo'`
-        - a GET argument called `'tzinfo'`
-        - the provided default timezone
-
-        """
-        tzinfo = self.tzinfo or \
-            (self.args and self.args.get('tzinfo')) or \
-            default
-        if isinstance(tzinfo, basestring):
-            tzinfo = timezone(tzinfo)
-        self.tzinfo = tzinfo
-        return self.tzinfo
-
-
-    def get_locale(self, default='en'):
-        """Returns the locale that should be used for this request as a
-        `babel.Locale` instance.
-
-        Tries the following in order:
-        - an attribute called `'locale'`
-        - a GET argument called `'locale'`
-        - the best guess from the ACCEPT_LANGUAGES header
-        - the language of the user agent
-        - the provided default language
-
-        """
-        locale = self.locale or \
-            (self.args and self.args.get('locale')) or \
-            self.accept_languages.best or \
-            self.user_agent.language or \
-            default
-        if isinstance(locale, basestring):
-            locale = locale.replace('-', '_')
-            locale = Locale.parse(locale, sep='_')
-        self.locale = locale
-        return self.locale
 
 
 class Response(BaseResponse):